refactor: rename sort_gitignore to sort_file_contents.
The convention of passing .gitignore as the file argument to a git hook results in some file containing the word gitignore in their names getting sorted as well. This changes avoids that while also reflecting the fact that the hook has utility beyond simply gitignore.
