+++ setup.cfg
@@ -29,7 +29,7 @@ install_requires =
 console_scripts =
     add-copyright = add_copyright_hook.add_copyright:main
     add-msg-issue = add_msg_issue_hook.add_msg_issue:main
-    sort-gitignore = sort_gitignore_hook.sort_gitignore:main
+    sort-file-contents = sort_file_contents_hook.sort_file_contents:main
 
 [options.packages.find]
 exclude =
